test(order): Order 통합 테스트 구현 및 lazy 로딩 버그 수정
- OrderIntegrationTest: 장바구니 주문·바로구매·조회·취소·ADMIN 상태 변경 16개 시나리오 통과
- OrderCreateRequest.quantity int → Integer (카트 주문 시 @Min 오탐 수정)
- OrderService.createDirect: SOLD_OUT/INACTIVE 상품 PRODUCT_NOT_AVAILABLE 검증 추가
- OrderRepository.findByIdWithItems: JOIN FETCH로 items lazy 로딩 문제 해결
- OrderService.getOrders: 목록 조회 시 items 강제 초기화
- OrderServiceTest: findByIdWithItems mock 반영

Closes #52

+++ backend/AgentCart/src/main/java/com/agentcart/order/controller/OrderController.java
@@ -40,7 +40,7 @@ public class OrderController {
                     request.getRecipientName(), request.getPhone(),
                     request.getAddress(), request.getAddressDetail());
         } else {
-            order = orderService.createDirect(memberId, request.getProductId(), request.getQuantity(),
+            order = orderService.createDirect(memberId, request.getProductId(), request.getQuantity().intValue(),
                     request.getRecipientName(), request.getPhone(),
                     request.getAddress(), request.getAddressDetail());
         }
